Implement scorlling BasicUIView grids to the bottom plus refactoring
Upon arrival of new records to the backing ObservableCollection, scroll the DataGrid to the last item. However, when the DataGrid is initialized, there's not yet a data context set, so when that happens, then we can listen to the CollectionChanged event to trigger scrolling (only really needed for additions).

In passing, also set the DataGrid to ReadOnly.

In the other two files, minor simplifications/cleanup.