Workaround journal bug: Already mapped bodies are not undiscovered (#85)
Just checking something is not already discovered is not enough. Many valuable bodies in the bubble mis-fire this criteria because the journal reports them as not discovered (because they have no first discoverer because they're well known systems). Checking that they're also not mapped ensures this doesn't happen.

Also simplified rendering of the string info.