[Framework] Make the new CoalescingId event arg nullable
While testing, I discovered that plugins that don't yet support sending the Coalescing ID were not mixing well because the default value of the ID is 0 -- which is also a valid body id. If the default is null, receivers can detect the absent value and apply a sensible default (ie. group them where it makes sense).