[Core] Visual improvements to notifications (#144)
* [Core] Visual improvements to notifications

- Notification title was previously missing; it's back!
- Fixed window and label backgrounds to be correctly/consistently transparent.

The click-thru bit is only partially working (transparent areas are click-thru).

* [Core] Reduce notification font size to avoid wrapping

Now that titles work, I noticed a few notification titles were wrapping. Examples:
- System Discovery Complete (from Evaluator)
- Minimum Distance Reached (from BioInsights)

To address, I made the following adjustments:
* Made the test notification title longer by adding the word "Popup" so it is now "Test Popup Notification".
* Reduced the Title font size to 20 (from 24).
* Moved the content box up to close the gap a bit.

* [Core] Exclude the notification popup colour picker button from themes

Noticed it was not the colour it should be and thought the setting was always being reset -- turns out the setting was correct, the button colour was not.

Here's a quick hack to fix.

* [Core] Further notification title tweaking

Now "System Discovery Complete" *actually* fits.

* [Core] Widen the native pop-up font selector so you can see variants

The previous width was a little bit narrow for the family of fonts like Segoe UI Variable...

The new width is 2x the original size and fits all fonts on my system and should do the trick for the moment.