[Core] Make button styles consistent throughout the app (#152)
Some core settings buttons and settings window buttons were inconsistent with the cluster of 4 buttons at the bottom of the Core window and below the Plugin list. I took those to be the defacto standard.

Applied the following changes to all inconsistent buttons (some programmatically for generated settings UI and the rest via designer):

```
button.FlatStyle = FlatStyle.Flat;
button.FlatAppearance.BorderSize = 0;
```

The only button I didn't touch was the Colour swatch button -- so if the selected colour is the same as the background for some reason, you can still see it.