feat(core,framework): Add LogMonitorState.BatchCancelled
We now have an explicit state that is fired when you cancel a read-all. This is useful for plugins which store state that may be incomplete/unusable if read all is not complete.

State transitions will be as follows: <initial state> -> <initial state> | Batch [-> <initial state> | BatchCancelled] -> <initial state>

The BatchCancelled state is not entered if the read-all run completes normally.