chore(framework,botanist): Add a no-op column sorter to Framework
There are likely a few other plugins (ie. any plugin which controls its own grid ordering) which likely want this too, so the idea is to make one many plugins can share to disable sorting in the new UI.

There was already such an implementation in Botanist; moved it to Framework and use it instead.