fix(core): Improve audio error handling, bug fixes and refactoring (#166)
* fix(herald): Lets try touching file info object INSIDE try/catch

And avoid breaking herald entirely...

* fix(herald): Also protect the instant playback path.

And early-out if the input file-path is an empty string or the file does not exist.

* fix(core): Add extra file checks and don't delete files after playing

Deleting files after playing resulted in race-condition playback failures and defeats the cache.

* fix,feat(framewwork,core): Add file cleanup option, refactor

After further discussion with the author of #160, the File.Delete was to cleanup files generated by Native -- which, of course, interfered with Herald's cache of files. However, native doesn't know when its file got played so never knows when to clean up it's file.

To solve this, I've added an `AudioOptions` parameter class to the PlayAudioFile method on the Core interface to provide playback options. Available options include `Instant` and `DeleteAfterPlay`, both are false by default.

These options are passed into the AudioHandler which acts on them accordingly.

NativeVoice now sets the `DeleteAfterPlay` option so its files are cleaned up by the AudioHandler. Herald does not set this and manages its own cache.

In the process, performed a few refactors:
* Replaced multiple optional boolean parameters to EnqueueAndPlay with the aformentioned parameter class.
* Factored out a PlayAudioFile method in AudioHandler to consolidate both instant and queued playback back into a single method.
* Widened some of the native Voice settings (Speed slider and Voice dropdown) because the voice names were cut off and we have plenty of space. (Also aligned other settings to match/be consistent.) As a result, merged #165 into this to avoid conflicts.
* Replaced the use of a bare list in AudioHandler with a ConcurrentDictionary because it was being used concurrently.