fix(core): Sort files by filename timestamp, mtime as fallback (#171)
If file modification time (mtime) is altered in any way (for example, restoring from backup or zip file in a way that does not preserve mtimes, etc.), the ordering of the journal files may become arbitrary and could affect plugin output or violate some assumptions. This parses the timestamp within the filename itself (taking into account both the old and new timestamp formats) as the authoritative timestamp, falling back to mtime if a timestamp could not be found in the filename.

Tested by adding blank lines to my first journal file. The previous mtime-based logic will now sort this as the most recent file instead of the first in the list. The updated logic introduced here keeps the modified file first in the list.