feat(core,botanist,explorer,herald): Add "About" info support
From an idea suggested on discord to a PR in one afternoon.

This adds a new form for showing "About {X}" where `{X}` is any plugin or core. The new form shows the "about" information (see below for what that includes) and there's now a "About Plugin" button below the list to open this view. (There's a plan/todo to create a toolbar or something eventually.) The plugin list now also includes an Author column.

The github and Donate links in the bottom left corner of the Core window have been replaced with a single "ℹ️" button (which has a tooltip to help understand it's purpose) which opens an About view for Core itself.

The about view shows the following info:
* Full name
* Short name
* Author name
* A multiline description
* Up to 4 links.

For plugins, this information is (optionally) provided via the IObservatoryPlugin.AboutInfo property which provides a new `AboutInfo` object. See the bundled plugin changes for a suggested implementation/usage.