fix(core): Fix signature checks, add support for signed 3P plugins
Fixes the signature checks for the bundled plugins so they're correctly detected as "Signed" if they are in fact signed with the same key as ObsCore.

Previously, no 3rd party plugin could be recognized as trusted -- even if signed with a trusted CA certificate. This is now supported.  (At least, self-signed plugins where I've trusted my own cert work.)

With previous logic (if signature checking was working), each version of an unsigned plugin file had to be accepted by the user once to run. Each time the plugin is updated (such that the file hash changes), it would need to be re-accepted -- which could be quite disruptive. With this change, plugins can be allowed by *signature thumbprint* so that multiple plugins from the same author require only 1 acceptance for all plugins and all versions which are signed by the same cert.

So for plugin authors with many plugins, it is now highly worthwhile setting up plugin signing.