fix(core): Targeted notifications now received by overriding notifiers
MattG reported that targeting ONLY a native notifier (ie. voice or visual) when there was an override in place was not received. We've never noticed that because apparently nobody's targeted *only* NativeVocal and there's no NativeVisual notifiers (yet). (It worked it PluginNotifier was also targeted.)

So yeah, turns out that was not handled correctly; and it's now fixed.

Tested with hacking up my Helm (targeted notification sender) and Aggregator (overriding notification receiver) plugins and Herald.