fix(core): Clamp numeric settings before assigning to Value
The NumericUpDown control Value property will throw if the value is out of bounds.

This can happen if a plugin sets bounds on a setting that does not include zero but does not initialize the value (typically when adding a new numeric value to a settings config that already exists).