fix(core): Minor FocusPlugin fix (#195)
Tabs use short name but windows use the full name. The original FocusPlugin implementation used the short name passed in from the calling plugin forwarded to it (likely) from a notification sender. This would only succeed in focusing a popped-out window if the Short name and Full name matched. There are plugins for which this is not true (ie. Commander vs. Fleet Commander).

Update PluginCore to find the plugin instance with the corresponding short name and use pass the plugin instance to the method which can then use either the full name to find the window or the short name to find the tab.