fix(explorer): Earth is not a primary star... (#200)
Many bodies in the bubble which have names that are not prefixed by the system name are being titled as "Primary star" incorrectly. ELWs and other named bodies in the bubble often fall victim to this bug (but it's probably not noticed because you don't generally need to FSS systems in the bubble and few people scan nav beacons with ObsCore running, where this would be exposed).