feat(framework,core): Add support for PP2.0 objects, StationService enum
Looked at recent changes to the jixxed/ed-journal-schemas and patching in PP2.0 and a colonisation change for a missing enum value which can cause some journals to fail to deserialize. There are a few new localized strings (ie. StationName_Localised -- which would be useful to check if populated).

Of note is the new CarrierLocation event (emitted at game startup) which is now collected and replayed by LogMonitor.

I suspect more colonisation journals will come once the feature is out of beta.