fix(framework): Add another enum value, fix it with generic converter
Discovered another enum value for StationService: ColonisationContribution

`StationServiceConverter` is now deprecated in favor of the generic one added in b944dc1 (which appears to work with flag-based enums too).

NOTE: I removed the redundant `Power` property from CarrierJump -- it's defined in FSDJump (base class) and was unintentionally hiding it. AFAIK, this should not break compatibility. Can revert if it does.