feat(framework,core): Add support for PP2.0 objects, StationService enum (#213)
* feat(framework,core): Add support for PP2.0 objects, StationService enum

Looked at recent changes to the jixxed/ed-journal-schemas and patching in PP2.0 and a colonisation change for a missing enum value which can cause some journals to fail to deserialize. There are a few new localized strings (ie. StationName_Localised -- which would be useful to check if populated).

Of note is the new CarrierLocation event (emitted at game startup) which is now collected and replayed by LogMonitor.

I suspect more colonisation journals will come once the feature is out of beta.

* fix(framework): Add another enum value, fix it with generic converter

Discovered another enum value for StationService: ColonisationContribution

`StationServiceConverter` is now deprecated in favor of the generic one added in b944dc1 (which appears to work with flag-based enums too).

NOTE: I removed the redundant `Power` property from CarrierJump -- it's defined in FSDJump (base class) and was unintentionally hiding it. AFAIK, this should not break compatibility. Can revert if it does.