fix(core): Fix message type cloning and fix filtering of self messages (#226)
Following up on 071f1da: This change resulted in the cloned message type becoming the string literal "System.Char[]" instead of the intended string. After some further poking around reworked to this to get what I believe was the original intent with the use of `Span` originally.

Also adds the filtering self-sent message implemented by @matt-g-dev in 0c907dd to the pure new-style messages (missed in 3c43fea).

Also, distinguish between the Legacy and modern On*PluginMessageEvent methods in the error report.

Finally, move the disabled plugin filter into a `Where` clause that provides the iterator of the `for` loop.