feat(core, framework): v2 Plugin update check and update folder
Implement a "v2" plugin update check which returns a PluginUpdateInfo object that allows for more fine-grained responses and URL customization. Marked the V1 version as deprecated. If the v2 returns no change, the v1 method is also checked.

Additionally, a new property has been added to Core that exposes a folder for writing plugin updates if plugins choose to auto-download their updates.